feat: Test VRF
