+++ contracts/test/v0.1/VRF-test-data/data.json
@@ -55,12 +55,12 @@
       },
       {
         "description": "too short",
-        "revert": "Malformed VRF proof",
+        "revert": "hex data is odd-length",
         "pi": "0x12345"
       },
       {
         "description": "too long",
-        "revert": "Malformed VRF proof",
+        "revert": "hex data is odd-length",
         "pi": "0x031f4dbca087a1972d04a07a779b7df1caa99e0f5db2aa21f3aecc4f9e10e85d0814faa89697b482daa377fb6b4a8b0191a65d34a6d90a8a2461e5db9205d4cf0bb4b2c31b5ef6997a585a9f1a72517b6f12345"
       }
     ]
@@ -121,12 +121,12 @@
       },
       {
         "description": "too short",
-        "revert": "Malformed compressed EC point",
+        "revert": "hex data is odd-length",
         "compressed": "0x12345"
       },
       {
         "description": "too long",
-        "revert": "Malformed compressed EC point",
+        "revert": "hex data is odd-length",
         "compressed": "0x03c2704fed5dc41d3979235b85edda8f86f1806c17ce0a516a034c605d2b4f9a2612345"
       }
     ]
